refactor: move STATUS and runIds to common to break layer violation
The agent layer was importing from @progressView (presentation layer),
creating an inappropriate dependency from business logic to UI code.

- Create src/common/constants/ with streamStatus.ts and runIds.ts
- Update agent layer files to import from @common/constants
- Make progressView re-export from common for backwards compatibility

This improves separation of concerns and allows the agent layer to be
used independently of the presentation layer.
