+++ src/common/constants/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Common constants shared across the codebase.
+ *
+ * This module provides a single source of truth for:
+ * - Stream status values
+ * - Run ID utilities
+ */
+
+export * from './streamStatus';
+export * from './runIds';
